+++ components/select/const.ts
@@ -1,4 +1,4 @@
-export const PROVIDE_KEY = Symbol('FSelect');
+export const SELECT_PROVIDE_KEY = Symbol('FSelect');
 
 export const PADDING_LEFT_BASE = 8;
 export const PADDING_LEFT_INDENT = 16;
